alterando

+++ backend/database/migrations/2020_06_14_044818_create_declarations_table.php
@@ -17,7 +17,7 @@ class CreateDeclarationsTable extends Migration
             $table->id('declarations_id');
             $table->date('date');
             $table->integer('type');
-            $table->integer('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->foreign('company_id')->references('id')->on('companies');
         });
     }
